feat: Atomizing the timer
