+++ src/components/organism/timer/index.jsx
@@ -1,29 +1,15 @@
-import {Text, View, StyleSheet } from "react-native"
+import React from 'react';
+import TimeDisplay from '../../atoms/time-display';
+import TimerContainer from '../../molecules/timer-container';
 
-export default function Timer({time}) {
-    const formattedTime = `${Math.floor(time / 60)
+export default function Timer({ time }) {
+  const formattedTime = `${Math.floor(time / 60)
     .toString()
     .padStart(2, "0")}:${(time % 60).toString().padStart(2, "0")}`;
-    
-    return (
-        <View style={styles.container}>
-            <Text style={styles.time}>{formattedTime}</Text>
-        </View>
-    );
-}
 
-const styles = StyleSheet.create({
-    container: {
-        flex: 0.3,
-        justifyContent: "center",
-        backgroundColor: "#F2F2F2",
-        padding: 15,
-        borderRadius: 15,
-    },
-    time: {
-        fontSize: 80, 
-        fontWeight: "bold",
-        textAlign: "center",
-        color:"#333333",
-    }
-})
+  return (
+    <TimerContainer>
+      <TimeDisplay formattedTime={formattedTime} />
+    </TimerContainer>
+  );
+}
